// src/types.ts
export interface DanmakuItem {
  time: number;
  mode: 1 | 4 | 5;
  fontSize: number;
  color: number;
  text: string;
  uid?: number;
  isAdmin?: boolean;
  badgeLevel?: number;
}